LayoutTests/js/dfg-inline-constructor-that-uses-arguments-expected.txt

Tests that we can inline a constructor that uses arguments without failing DFG validation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS bar().x is 42
PASS successfullyParsed is true

TEST COMPLETE
